Initial commit
`hop` is a shell-script utility designed for downloading Advent of Code
puzzles and inputs, as well as submitting answers for the puzzles
directly from the command line.

The name abbreviation is from words *h*oliday *o*f *p*uzzles!

`hop` works on top of aoc-cli and provides two custom commands puzzle
and answer. Full functionality of the aoc-cli can be accessed through
hop if puzzle or answer commands are not used.

The puzzle command fetches the current or specific puzzle desscription
and input.

The answer command submits an answer for the current of specific puzzle.

The session token is managed automatically if user is using MacOS and
firefox. The session token is queried from the cookies database.